requirements.txt
matplotlib==3.5.3
numpy==1.21.5
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
Pillow==9.2.0
scikit_learn==1.0.2
scipy==1.7.3
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
